Create README.md
My name is Ogbechie Judith, and I am a passionate software developer with a keen interest in Python programming. With a strong background in Python, I am dedicated to leveraging its power to create efficient and robust software solutions. I strive to stay up-to-date with the latest advancements in Python and continuously expand my knowledge to tackle complex challenges. I am motivated by the satisfaction of turning ideas into functional and user-friendly applications. With a detail-oriented approach and a commitment to delivering high-quality code, I am enthusiastic about contributing to innovative projects and collaborating with like-minded professionals.